Setting up the state machine basics for testing.
Setting up a runner for the state machine, and stubbing out the testing
api and server to simply be the main api and server for now.

Setting up initial types to hold the model state, but these will have to
change soon as they don't include any authenitication values at the
moment.