Expanding README, cleaning up API, and extending tests
Expanding the README with more information about my choices and how the
code is structured, plus some operation information.

Cleaning up the API, as it was a bit messy, and if slightly modified it
would be more presentable and explainable.

Expanding tests, they are still failing because the generation isn't
constrained enough to make sure it is only using things that exist in
the model state, rather than something that _was_ but has since been
pruned from the generation tree.